# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Express — Hợp đồng tích hợp backend",
    "version": "0.2.0",
    "description": "Frontend HTTP-only ngày 02/10/2026. Các endpoint cũ giữ nguyên. Register tạo shop rồi yêu cầu đăng nhập; không tự tạo session. Cookie và CSRF cần backend triển khai."
  },
  "servers": [
    {
      "url": "https://localhost:7001/api/v1",
      "description": "Ví dụ; thay bằng URL backend thực tế"
    }
  ],
  "security": [
    {
      "cookieSession": []
    }
  ],
  "paths": {
...
    "/pickup-locations": {
      "get": {
        "operationId": "listPickupLocations",
        "summary": "Danh sách điểm lấy hàng của shop trong phiên.",
        "description": "Quyền đề xuất: Shop. Hợp đồng tích hợp suy ra từ frontend, chưa xác nhận backend đã triển khai.",
        "tags": [
          "pickup-locations"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Thành công; trả JSON trực tiếp theo schema.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickupLocation"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập/hết phiên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trong phạm vi được phép",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, báo giá, năng lực hoặc kỳ đối soát",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createPickupLocation",
        "summary": "Tạo điểm lấy hàng; tên duy nhất trong từng shop.",
        "description": "Quyền đề xuất: Shop. Hợp đồng tích hợp suy ra từ frontend, chưa xác nhận backend đã triển khai.",
        "tags": [
          "pickup-locations"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Thành công; trả JSON trực tiếp theo schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupLocation"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập/hết phiên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trong phạm vi được phép",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, báo giá, năng lực hoặc kỳ đối soát",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickupLocationInput"
              },
              "example": {
                "name": "Cửa hàng chính",
                "contactName": "Shop Minh Họa",
                "contactPhone": "0900000001",
                "address": "Địa chỉ mẫu, TP. Hồ Chí Minh"
              }
            }
          }
        }
      }
    },
...
    "/shipments/{shipmentId}": {
      "get": {
        "operationId": "getShipment",
        "summary": "Chi tiết theo mã vận đơn; phải kiểm tra quyền từng đơn, không phải tra cứu công khai.",
        "description": "Quyền đề xuất: Shop/Staff/Operations. Hợp đồng tích hợp suy ra từ frontend, chưa xác nhận backend đã triển khai.",
        "tags": [
          "shipments"
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "shipmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID dạng chuỗi; dùng ID thực từ API, không giả định chỉ là UUID."
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công; trả JSON trực tiếp theo schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập/hết phiên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trong phạm vi được phép",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, báo giá, năng lực hoặc kỳ đối soát",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Thông báo lỗi để hiển thị cho người dùng"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PickupLocation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "contactName",
          "contactPhone",
          "address"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "PickupLocationInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "contactName",
          "contactPhone",
          "address"
        ]
      },
...
    },
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Express.Session",
        "description": "Tên cookie đề xuất; dùng HttpOnly server session. Không phải JWT hiện có."
      }
    }
  }
}